+++ test/unit/spec_helper.rb
@@ -1,7 +1,11 @@
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+$LOAD_PATH.unshift File.expand_path('../../../lib', __FILE__)
+
+require 'rspec/core'
 require 'vagrant-omnibus'
 
 RSpec.configure do |config|
+  config.formatter = :documentation
+
   # a little syntactic sugar
   config.alias_it_should_behave_like_to :it_has_behavior, 'has behavior:'